[FEAT/#71] 위시 패키지 정리
