+++ app/src/main/java/com/example/teumteum/data/entities/Wish.kt
@@ -1,5 +1,6 @@
 package com.example.teumteum.data.entities
 
+import com.example.teumteum.data.remote.wish.WishCategory
 import com.google.gson.annotations.SerializedName
 
 data class WishItem(
@@ -14,5 +15,5 @@ data class Wish(
     @SerializedName(value = "id") val id: Long,
     @SerializedName(value = "title") val title: String,
     @SerializedName(value = "estimatedDuration") val estimatedDuration: String,
-    @SerializedName(value = "categories") val categories: List<Category>
+    @SerializedName(value = "categories") val categories: List<WishCategory>
 )
