+++ app/src/main/java/com/example/teumteum/data/remote/wish/WishCategory.kt
@@ -2,7 +2,7 @@ package com.example.teumteum.data.remote.wish
 
 import com.google.gson.annotations.SerializedName
 
-data class Category(
+data class WishCategory(
     @SerializedName("id") val id: Long,
     @SerializedName("name") val name: String
 )
