+++ app/src/main/java/com/example/teumteum/data/remote/wish/WishResponse.kt
@@ -1,4 +1,4 @@
-package com.example.teumteum.data.remote
+package com.example.teumteum.data.remote.wish
 
 import com.example.teumteum.data.entities.Wish
 import com.google.gson.annotations.SerializedName
