+++ app/src/main/java/com/example/teumteum/ui/wish/WishRegisterFragment.kt
@@ -9,10 +9,11 @@ import android.view.ViewGroup
 import android.widget.Toast
 import androidx.core.content.ContextCompat
 import com.example.teumteum.R
-import com.example.teumteum.data.remote.RegisterWishRequest
-import com.example.teumteum.data.remote.WishService
+import com.example.teumteum.data.remote.wish.RegisterWishRequest
+import com.example.teumteum.data.remote.wish.WishService
 import com.example.teumteum.databinding.FragmentWishRegisterBinding
 import com.example.teumteum.ui.todo.TodoRegisterFragment
+import com.example.teumteum.ui.wish.view.RegisterWishView
 import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.google.android.material.bottomsheet.BottomSheetDialog
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
