+++ app/src/main/java/com/example/teumteum/ui/wish/WishlistEditFragment.kt
@@ -9,6 +9,7 @@ import androidx.fragment.app.Fragment
 import com.example.teumteum.R
 import com.example.teumteum.data.entities.WishItem
 import com.example.teumteum.databinding.FragmentWishlistEditBinding
+import com.example.teumteum.ui.wish.adapter.WishlistEditRVAdapter
 import com.google.android.material.bottomnavigation.BottomNavigationView
 
 class WishlistEditFragment(private val wishlist: MutableList<WishItem>) : Fragment() {
