+++ app/src/main/java/com/example/teumteum/ui/wish/adapter/WishlistRVAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.teumteum.ui.wish
+package com.example.teumteum.ui.wish.adapter
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,6 +8,8 @@ import androidx.recyclerview.widget.RecyclerView
 import com.example.teumteum.R
 import com.example.teumteum.data.entities.WishlistItem
 import com.example.teumteum.databinding.ItemWishlistBinding
+import com.example.teumteum.ui.wish.WishEditFragment
+import com.example.teumteum.ui.wish.WishSetting01Fragment
 
 class WishlistRVAdapter(private var wishlist: List<WishlistItem>, private val fragmentManager: FragmentManager) : RecyclerView.Adapter<WishlistRVAdapter.ViewHolder>() {
