+++ app/src/main/java/com/example/teumteum/ui/wish/view/RegisterWishView.kt
@@ -1,0 +1,6 @@
+package com.example.teumteum.ui.wish.view
+
+interface RegisterWishView {
+    fun onRegisterSuccess(code: String)
+    fun onRegisterFailure(code: String)
+}
